⚡️ [#63] 빠진 기능들 추가
