+++ DoriDori_iOS/DoriDori_iOS/Common/LocationManager/LocationManager.swift
@@ -9,20 +9,6 @@ import Foundation
 import CoreLocation
 import RxSwift
 
-enum DoriDoriLocationError: LocalizedError {
-    case noData
-    case unknownError
-    case timeout
-    
-    var errorDescription: String? {
-        switch self {
-        case .noData: return "위도,경도 정보가 없습니다"
-        case .unknownError: return "위치를 불러오는데 알 수 없는 에러가 발생했습니다"
-        case .timeout: return "위치를 불러오는데 시간이 초과되었습니다."
-        }
-    }
-}
-
 final class LocationManager: NSObject {
     private let location: CLLocationManager
     private var completionHander: ((Result<Location, DoriDoriLocationError>) -> Void)?
@@ -32,7 +18,6 @@ final class LocationManager: NSObject {
     ) {
         self.location = location
         super.init()
-//        self.location.delegate = self
         self.location.desiredAccuracy = kCLLocationAccuracyBest
     }
     
@@ -56,10 +41,8 @@ final class LocationManager: NSObject {
         
         switch status {
         case .authorizedAlways, .authorizedWhenInUse:
-            print("🍀위치정보 허용되어있음")
             self.location.startUpdatingLocation()
         case .denied, .notDetermined:
-            print("🍀위치정보 허용안되어있음")
             self.location.requestWhenInUseAuthorization()
             self.location.startUpdatingLocation()
         default: break
@@ -68,36 +51,4 @@ final class LocationManager: NSObject {
         return Observable.merge(locationOb, errorOb)
     }
     
-}
-
-// MARK: - CLLocationManagerDelegate
-
-extension LocationManager: CLLocationManagerDelegate {
-    
-    func locationManager(
-        _ manager: CLLocationManager,
-        didFailWithError error: Error
-    ) {
-        self.completionHander?(.failure(DoriDoriLocationError.unknownError))
-    }
-
-    func locationManager(
-        _ manager: CLLocationManager,
-        didUpdateLocations locations: [CLLocation]
-    ) {
-        if locations.isEmpty {
-            self.completionHander?(.failure(DoriDoriLocationError.noData))
-            return
-        }
-        guard let currentCoordinator = locations.last?.coordinate else { return }
-        let location: Location = (longitude: currentCoordinator.longitude, latitude: currentCoordinator.latitude)
-        self.completionHander?(.success(location))
-    }
-    
-    func locationManager(
-        _ manager: CLLocationManager,
-        didFinishDeferredUpdatesWithError error: Error?
-    ) {
-        self.completionHander?(.failure(DoriDoriLocationError.timeout))
-    }
 }
